+++ CHANGELOG.md
@@ -13,6 +13,10 @@ https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Changed
+
+- Move all devices initialization to `all_devices.py`
+
 ## [0.1.0] - 2025-04-09
 
 ### Added